pyproject.toml
[tool.poetry]
name = "seatlib"
version = "0.2.0"
description = "library seat watcher for Tsinghua"
authors = ["Bryan Lai <bryanlais@gmail.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
include = ["*.yml"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
confuse = "^2.0"
pytest = "^7.3"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
seatlib = 'seatlib:execute'

[tool.ruff]
select = ["E", "F", "W"]
ignore = [
    # "F403",
    # "F405",
    "E402", # imports not at the very top
]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
